gdk: Convert mouse position to doubles, add new getters
We've long had double precision mouse coordinates on wayland (e.g.
when rotating a window) but with the new scaling we even have it on
X (and, its also in Xinput2), so convert all the internal mouse/device
position getters to use doubles and add new accessors for the
public APIs that take doubles instead of ints.

+++ gdk/gdkinternals.h
@@ -358,7 +358,7 @@ void _gdk_window_invalidate_for_expose (GdkWindow       *window,
                                         cairo_region_t       *region);
 
 GdkWindow * _gdk_window_find_child_at (GdkWindow *window,
-                                       int x, int y);
+                                       double x, double y);
 GdkWindow * _gdk_window_find_descendant_at (GdkWindow *toplevel,
                                             double x, double y,
                                             double *found_x,
@@ -380,8 +380,8 @@ void _gdk_synthesize_crossing_events (GdkDisplay                 *display,
                                       GdkDevice                  *device,
                                       GdkDevice                  *source_device,
 				      GdkCrossingMode             mode,
-				      gint                        toplevel_x,
-				      gint                        toplevel_y,
+				      gdouble                     toplevel_x,
+				      gdouble                     toplevel_y,
 				      GdkModifierType             mask,
 				      guint32                     time_,
 				      GdkEvent                   *event_in_queue,
